# docs/api-docs.yaml
openapi: 3.0.1
info:
  title: Tasa-API
  description: "API for synchronize TASA app data."
  version: "1.0"
servers:
  - url: http://localhost:8080/api
...
paths:
...
  /event/create:
    post:
      security:
        - bearerAuth: [ ]
      tags:
        - Events
      summary: Creates a event
      description: Register a event of a user given it's event identifier and calendar identifier.
      requestBody:
        description: Event data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventOutput'
        '400':
          description: Event name is empty.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: A event with given event id and calendar id already exists.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /event/{id}/update/title:
    put:
      security:
        - bearerAuth: []
      tags:
        - Events
      summary: Update a event
      description: Updated the title of a event.
      parameters:
        - name: id
          in: path
          description: Event identifier
          required: true
          schema:
            type: integer
            format: int32
            example: 5287
      requestBody:
        description: Event data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
        required: true
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventOutput'
        '400':
          description: Event name is empty.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Event not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /event/{id}:
    get:
      security:
        - bearerAuth: []
      tags:
        - Events
      summary: Retrieves a event
      description: Get a event given its id and respetive calendar id.
      parameters:
        - name: id
          in: path
          description: Event identifier
          required: true
          schema:
            type: integer
            format: int32
            example: 5287
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventOutput'
        '400':
          description: Event identifier is negative.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Event not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /event/all:
    get:
      security:
        - bearerAuth: []
      tags:
        - Events
      summary: Retrieves all events of a user
      description: Get all the registerd events of an user.
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventList'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /event/remove/{id}:
    delete:
      security:
        - bearerAuth: []
      tags:
        - Events
      summary: Remove a event of a user
      description: Deletes a event given its identifier and calendar identifier.
      parameters:
        - name: id
          in: path
          description: Event identifier
          required: true
          schema:
            type: integer
            format: int32
            example: 5208
      responses:
        '200':
          description: Ok
        '400':
          description: Event identifier is negative.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Event not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
  /rule/event/{id}:
    get:
      security:
        - bearerAuth: []
      tags:
        - Rules
      summary: Gets a rule event
      description: Gets a rule event by it's identifier.
      parameters:
        - name: id
          in: path
          description: Rule identifier
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleEventOutput'
        '400':
          description: Identifier is negative.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Rule not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

    delete:
      security:
        - bearerAuth: []
      tags:
        - Rules
      summary: Deletes a rule event
      description: Deletes a rule event by it's identifier.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Ok
        '400':
          description: Identifier is negative.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Rule not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    EventInput:
      type: object
      required:
        - title
        - endTime
        - startTime
      properties:
        startTime:
          type: string
          format: date-time
          example: "2024-10-01T12:00:00Z"
        endTime:
          type: string
          format: date-time
          example: "2024-10-01T12:00:00Z"
        title:
          type: string
          description: Event title
          example: Meeting.
    EventOutput:
      type: object
      required:
        - id
        - title
        - endTime
        - startTime
      properties:
        id:
          type: integer
          format: int32
          description: Event identifier
          example: 1025
        startTime:
          type: string
          format: date-time
          example: "2024-10-01T12:00:00Z"
        endTime:
          type: string
          format: date-time
          example: "2024-10-01T12:00:00Z"
        title:
          type: string
          description: Event title
          example: Meeting.
    EventList:
      type: array
      properties:
        nEvents:
          type: integer
          description: number of results
          example: 20
        events:
          type: array
          items:
            $ref: '#/components/schemas/EventOutput'
...
    RuleEventOutput:
      type: object
      required:
        - id
        - endTime
        - startTime
        - event
      properties:
        id:
          type: integer
          description: Rule dentifier
          example: 10
        startTime:
          type: string
          format: date-time
          description: Start time of rule
          example: "2024-10-01T12:00:00Z"
        endTime:
          type: string
          format: date-time
          description: End time of rule
          example: "2024-10-01T12:00:00Z"
        event:
          $ref: '#/components/schemas/EventOutput'
...
    Problem:
      description: Error response
      type: object
      properties:
        type:
          type: string
        title:
          type: string
          format: uri
        status:
          type: integer
        detail:
          type: string